v0.1.0
commit a21edbe78354b0544766bdff7b8b4b0df530fd32
Author: Hubert Lin <linhub15@gmail.com>
Date:   Thu Nov 1 10:39:46 2018 -0600

    app accepts :5001 url for dev and relative for prod

commit 844c0cdd1ed7983046adf816511ae41a216e01f5
Author: Hubert Lin <linhub15@gmail.com>
Date:   Wed Oct 31 16:38:38 2018 -0600

    environment variables for base url

commit 732cb4f99df6c3a87c4eac861ed3bbbb89a6777f
Author: Hubert Lin <linhub15@gmail.com>
Date:   Wed Oct 31 15:07:46 2018 -0600

    correct dist folder path

commit c521dc431a0689db0b41764a0287bf6b1a026075
Author: Hubert Lin <linhub15@gmail.com>
Date:   Wed Oct 31 13:25:41 2018 -0600

    prevent illegal white spaces employee input

commit 62a61a9aefecb5d6a437cdab04ebe3896be1acd8
Author: Hubert Lin <linhub15@gmail.com>
Date:   Wed Oct 31 11:15:59 2018 -0600

    add loading spinner to original app load

commit 65e558e7521a09a58736992971f58c3e9a6f5b22
Author: Hubert Lin <linhub15@gmail.com>
Date:   Wed Oct 31 11:08:04 2018 -0600

    gracefully deal with page not found

commit 1160fde5dafcb414bd6c3dbc9ef706e51a23fd25
Author: Hubert Lin <linhub15@gmail.com>
Date:   Wed Oct 31 10:53:41 2018 -0600

    Fix navigation issue

    only the text caused navigation

commit c5acfab4968a42dd956c8c5fc4eccf8523a874d0
Author: Hubert Lin <linhub15@gmail.com>
Date:   Wed Oct 31 10:47:24 2018 -0600

    gracefully deals with empty timesheets

    http handles errors

commit 09b45c69656aa854d54f8e2c571d15c86a07cd5a
Author: Hubert Lin <linhub15@gmail.com>
Date:   Tue Oct 30 16:26:14 2018 -0600

    fix hours not displaying on submit

commit 1f918168cb31146f48139368500f27289761a3ab
Author: Hubert Lin <linhub15@gmail.com>
Date:   Tue Oct 30 16:01:55 2018 -0600

    consistent styling

commit a004e3546719194a693a6700d03fe3c862bff5bf
Author: Hubert Lin <linhub15@gmail.com>
Date:   Tue Oct 30 14:53:24 2018 -0600

    consistent mat-card-list styling

commit 92ba9250015896d8dfbaaf9b2e6dca43e2bdf25f
Author: Hubert Lin <linhub15@gmail.com>
Date:   Tue Oct 30 14:00:29 2018 -0600

    Warning dialog for all HTTP DELETE
      warning dialog - Submit Time Sheet
      warning dialog - Delete Time Sheet from dialog
      warning dialog - Delete Time Sheet from list
      warning dialog - Delete Employee

commit 35d1fe81bb4b131e6378b0a597026f045a4c1a77
Author: Hubert Lin <linhub15@gmail.com>
Date:   Tue Oct 30 09:21:46 2018 -0600

    consolidate Snackbar to service for default duration

+++ src/app/employees/employee-list/employee-list.component.ts
@@ -1,20 +1,22 @@
 import { Component, OnInit } from '@angular/core';
-import { MatDialog, MatSnackBar } from '@angular/material';
+import { MatDialog } from '@angular/material';
 
 import { Employee, EmployeeService } from '../../core';
 import { EmployeeDialogComponent } from '../employee-dialog/employee-dialog.component';
+import { SnackBarService } from '../../shared';
+import { WarningDialogComponent } from '../../shared';
 
 @Component({
   selector: 'app-employee-list',
   templateUrl: './employee-list.component.html',
-  styleUrls: ['./employee-list.component.css']
+  styleUrls: ['./employee-list.component.scss']
 })
 export class EmployeeListComponent implements OnInit {
    employees: Employee[];
   constructor(
     private employeeService:EmployeeService,
     public dialog: MatDialog,
-    public snackBar: MatSnackBar) { }
+    public snackBar: SnackBarService) { }
 
   ngOnInit() {
     this.employeeService.getEmployees()
@@ -36,7 +38,7 @@ export class EmployeeListComponent implements OnInit {
       if (!employee) { return }
       this.employeeService.add(employee)
         .subscribe(e => this.employees.push(e))
-        this.snackBar.open('Added','', {duration: 2500});
+        this.snackBar.open('Added!');
     })
   }
   openEditDialog(employee: Employee): void {
@@ -54,15 +56,26 @@ export class EmployeeListComponent implements OnInit {
       this.employeeService.update(dialogEmployee).subscribe();
       let index = this.employees.indexOf(employee);
       this.employees[index] = dialogEmployee;
-      this.snackBar.open('Saved','', {duration: 2500});
+      this.snackBar.open('Saved!');
     });
   }
 
   deleteEmployee(employee: Employee): void {
     if (!employee) { console.log("required param"); return; }
-    this.employeeService.delete(employee.id).subscribe();
-    const index = this.employees.indexOf(employee);
-    this.employees.splice(index, 1)
-    this.snackBar.open('Deleted','', {duration: 2500});
+    const dialogRef = this.dialog.open(WarningDialogComponent, {
+      autoFocus: false,
+      data: {
+        title: 'Are you sure?',
+        message: 'Deleting '+employee.fullName()+' will permanently delete all their time sheets too.', 
+        submitAction: 'Delete'}
+    });
+
+    dialogRef.afterClosed().subscribe(submitted => {
+      if (!submitted) {return}
+      this.employeeService.delete(employee.id).subscribe();
+      const index = this.employees.indexOf(employee);
+      this.employees.splice(index, 1)
+      this.snackBar.open('Deleted ' + employee.fullName() + '!');
+    })
   }
 }
